# Use modern RNG