@ilan-gold 2024-05-02
@RubenVanEsch could you look at this as well for your [issue](https://github.com/scverse/scanpy/issues/2969)?

I am not so sure if a 15% slowdown is acceptable given the fact that the answer to solving this could just be "use WSL," which we never got a response on because we go sidetracked.  Could @RubenVanEsch or someone else confirm or not that the original problem is reproducible on WSL?