@RubenVanEsch 2024-05-02
@ilan-gold turns out i cant install WSL on my laptop after all, so unfortunately i cant check this.