@patrick-nicodemus 2024-09-05
Hi, I cloned this repo, switched to `modern-rng`, and installed it with `pip`. I was able to reproduce the same error.
```
Exception ignored in: <class 'ValueError'>
Traceback (most recent call last):
    File "numpy\random\_generator.pyx", line 622, in numpy.random._generator.Generator.integers
    File "numpy\random\_bounded_integers.pyx", line 2881, in numpy.random._bounded_integers._rand_int32"
ValueError: high is out of bounds for int32
```
I am using numpy 1.26, which is the numpy version required by this branch.